// fe-experiment/backend-for-ved/src/lib/enums/models/comment.enums.ts
export enum CommentEntityType {
  FORM_PAYMENT = 'FormPayment',
}
export enum CommentKind {
  INTERNAL = 'internal', // Внутренний (для сотрудников)
  EXTERNAL = 'external', // Внешний (для клиента)
}

export enum CommentPattern {
  SEND_NOTIFICATION = 'fea360.comment.send.notification',
}